requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
colorama==0.4.5
countryinfo==0.1.2
deprecation==2.1.0
dj-database-url==1.0.0
Django==4.0.4
django-admin-interface==0.19.1
django-admin-logs==1.0.2
django-ckeditor==6.4.2
django-cloudinary-storage==0.3.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-debug-toolbar==3.5.0
django-email-verification==0.3.1
django-extensions==3.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==2.0.0
django-json-ld==0.0.5
django-pandas==0.6.6
django-phone-field==1.8.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-api-logger==1.1.10
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
phone-iso3166==0.4.0
phonenumbers==8.12.51
Pillow==9.2.0
plotly==5.10.0
psycopg2==2.9.3
pycountry==22.3.5
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
tornado==6.2
twilio==7.13.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validate-email==1.3
webencodings==0.5.1
websocket==0.2.1
whitenoise==6.2.0
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
